backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0         # Async SQLite driver (zero install — built into Python)
alembic==1.13.1           # DB migrations

# ── Auth / Security ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT encode/decode
passlib[bcrypt]==1.7.4             # Password hashing
bcrypt==3.2.2                      # passlib 1.7.4 is incompatible with bcrypt>=4.0.0
python-multipart==0.0.9            # Form data (login endpoint)

# ── AI Integration ────────────────────────────────────────────────────────────
openai==1.30.5            # OpenAI Python SDK (v1+)
anthropic==0.26.0         # Optional: Anthropic Claude SDK

# ── Config & Validation ───────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9            # Starlette-native rate limiter (wraps limits)

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0             # Async HTTP (used internally by openai SDK)

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0             # Test client for FastAPI
